Revamped the whole Database Structure

+++ database/migrations/2025_08_21_135300_create_assessment_weights_table.php
@@ -18,7 +18,7 @@ class CreateAssessmentWeightsTable extends Migration
                 $table->increments('id');
                 $table->unsignedInteger('module_id');
                 $table->unsignedInteger('academic_year_id');
-                $table->unsignedInteger('assessment_type_id');
+                $table->unsignedBigInteger('assessment_type_id');
                 $table->string('description')->nullable();
                 $table->decimal('weight', 5, 2);
                 $table->timestamps();
